sidePanel API fixes
