+++ src/contentScript/sidebarController.tsx
@@ -40,19 +40,56 @@ import { getFormPanelSidebarEntries } from "@/contentScript/ephemeralFormProtoco
 import * as sidePanel from "@/sidebar/sidePanel/messenger/api";
 import { memoizeUntilSettled } from "@/utils/promiseUtils";
 import { getTimedSequence } from "@/types/helpers";
-import { backgroundTarget, getMethod } from "webext-messenger";
+import { backgroundTarget, getMethod, messenger } from "webext-messenger";
 import { isMV3 } from "@/mv3/api";
 
 const HIDE_SIDEBAR_EVENT_NAME = "pixiebrix:hideSidebar";
 
+/* Approximate sidebar width in pixels. Used to determine whether it's open */
+const MINIMUM_SIDEBAR_WIDTH = 300;
+
 export const isSidePanelOpen = isMV3()
-  ? sidePanel.isSidePanelOpen
+  ? isSidePanelOpenMv3
   : sidebarMv2.isSidebarFrameVisible;
 
-// eslint-disable-next-line local-rules/persistBackgroundData -- Function
-const isSidePanelOpenSync = isMV3()
-  ? sidePanel.isSidePanelOpenSync
-  : sidebarMv2.isSidebarFrameVisible;
+/**
+ * Determines whether the sidebar is open.
+ * @returns false when it's definitely closed
+ * @returns 'unknown' when it cannot be determined, beause the extra padding might be
+ *          caused by the dev tools being open on the side or due to another sidebar
+ */
+// The type cannot be `undefined` due to strictNullChecks
+function isSidePanelOpenSync(): false | "unknown" {
+  if (!isMV3()) {
+    throw new Error("isSidePanelOpenSync is only available in MV3");
+  }
+
+  if (!globalThis.window) {
+    return "unknown";
+  }
+
+  return window.outerWidth - window.innerWidth > MINIMUM_SIDEBAR_WIDTH
+    ? "unknown"
+    : false;
+}
+
+// This method is exclusive to the content script, don't export it
+async function isSidePanelOpenMv3(): Promise<boolean> {
+  if (isSidePanelOpenSync() === false) {
+    return false;
+  }
+
+  try {
+    await messenger(
+      "SIDEBAR_PING",
+      { retry: false },
+      await sidePanel.getSidebarTargetForCurrentTab(),
+    );
+    return true;
+  } catch {
+    return false;
+  }
+}
 
 // - Only start one ping at a time
 // - Limit to one request every second (if the user closes the sidebar that quickly, we likely see those errors anyway)
